+++ pkgs/ffigen/lib/src/code_generator/objc_methods.dart
@@ -19,6 +19,7 @@ import 'pointer.dart';
 import 'scope.dart';
 import 'type.dart';
 import 'typealias.dart';
+import 'union.dart';
 import 'utils.dart';
 import 'writer.dart';
 
@@ -621,12 +622,15 @@ class ObjCMethod extends AstNode with HasLocalScope {
         msgSendParams,
         structRetPtr: ptrVar,
       );
+      final compoundKind = returnType.typealiasType is Union
+          ? 'Union'
+          : 'Struct';
       s.write('''
     final $ptrVar = $calloc<$returnTypeStr>();
     $invoke;
     final $finalizableVar = $ptrVar.cast<$uint8Type>().asTypedList(
         $sizeOf<$returnTypeStr>(), finalizer: $calloc.nativeFree);
-    return ${context.libs.prefix(ffiImport)}.Struct.create<$returnTypeStr>(
+    return ${context.libs.prefix(ffiImport)}.$compoundKind.create<$returnTypeStr>(
         $finalizableVar);
 ''');
     } else {