+++ pkgs/ffigen/lib/src/code_generator/union.dart
@@ -41,6 +41,7 @@ class Union extends Compound {
     required super.context,
     super.nativeType,
     super.apiAvailability,
+    super.sizeInBytes,
   });
 
   @override