+++ pkgs/native_toolchain_c/test/native_toolchain/xcode_test.dart
@@ -11,6 +11,7 @@ import 'dart:io';
 import 'package:native_toolchain_c/src/native_toolchain/xcode.dart';
 import 'package:native_toolchain_c/src/tool/tool.dart';
 import 'package:native_toolchain_c/src/tool/tool_instance.dart';
+import 'package:process/process.dart';
 import 'package:test/test.dart';
 
 import '../helpers.dart';
@@ -59,6 +60,7 @@ void main() {
       sdk: 'doesnotexist',
       tool: tool,
       logger: logger,
+      processManager: const LocalProcessManager(),
     );
     expect(result, <ToolInstance>[]);
   });