+++ pkgs/ffigen/CHANGELOG.md
@@ -9,6 +9,8 @@
   - Rename `Headers` to `Input`
   - Remove `libraryImports`, which was dead code
   - Remove `useSupportedTypedefs`, treating it as always true
+- Fix [a bug](https://github.com/dart-lang/native/issues/3504) in handling of
+  small structs in ObjC on mac/iOS x64.
 - Minor Objective-C code generator and function type signature fixes.
 - Bump `package:code_assets` dependency to `^2.0.0`.